chore: rename the error handling page and update content
